Consolidate 4/7 new example clients down to a unified java benchmark style.
