--- examples/adperformance/client/adperformance/BaseBenchmark.java
@@ -1,299 +1,0 @@
-/* This file is part of VoltDB.
- * Copyright (C) 2008-2016 VoltDB Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
- * OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package adperformance;
-
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.concurrent.CountDownLatch;
-
-import org.voltdb.client.Client;
-import org.voltdb.client.ClientConfig;
-import org.voltdb.client.ClientFactory;
-import org.voltdb.client.ClientStats;
-import org.voltdb.client.ClientStatsContext;
-import org.voltdb.client.ClientStatusListenerExt;
-
-//import voter.procedures.Vote;
-
-public abstract class BaseBenchmark {
-
-    // handy, rather than typing this out several times
-    public static final String HORIZONTAL_RULE =
-            "----------" + "----------" + "----------" + "----------" +
-            "----------" + "----------" + "----------" + "----------" + "\n";
-
-    // validated command line configuration
-    final BenchmarkConfig config;
-    // Reference to the database connection we will use
-    final Client client;
-    // Timer for periodic stats printing
-    Timer timer;
-    // Benchmark start time
-    long benchmarkStartTS;
-    // Statistics manager objects from the client
-    final ClientStatsContext periodicStatsContext;
-    final ClientStatsContext fullStatsContext;
-
-    /**
-     * Provides a callback to be notified on node failure.
-     * This example only logs the event.
-     */
-    class StatusListener extends ClientStatusListenerExt {
-        @Override
-        public void connectionLost(String hostname, int port, int connectionsLeft, DisconnectCause cause) {
-            // if the benchmark is still active
-            if ((System.currentTimeMillis() - benchmarkStartTS) < (config.duration * 1000)) {
-                System.err.printf("Connection to %s:%d was lost.\n", hostname, port);
-            }
-        }
-    }
-
-    /**
-     * Constructor for benchmark instance.
-     * Configures VoltDB client and prints configuration.
-     *
-     * @param config Parsed & validated CLI options.
-     */
-    public BaseBenchmark(BenchmarkConfig config) {
-        this.config = config;
-
-        ClientConfig clientConfig = new ClientConfig(config.user, config.password, new StatusListener());
-        if (config.autotune) {
-            clientConfig.enableAutoTune();
-            clientConfig.setAutoTuneTargetInternalLatency(config.latencytarget);
-        }
-        else {
-            clientConfig.setMaxTransactionsPerSecond(config.ratelimit);
-        }
-        client = ClientFactory.createClient(clientConfig);
-
-        periodicStatsContext = client.createStatsContext();
-        fullStatsContext = client.createStatsContext();
-
-    printHeading("Command Line Configuration");
-        System.out.println(config.getConfigDumpString());
-    }
-
-    /**
-     * Connect to a single server with retry. Limited exponential backoff.
-     * No timeout. This will run until the process is killed if it's not
-     * able to connect.
-     *
-     * @param server hostname:port or just hostname (hostname can be ip).
-     */
-    void connectToOneServerWithRetry(String server) {
-        int sleep = 1000;
-        while (true) {
-            try {
-                client.createConnection(server);
-                break;
-            }
-            catch (Exception e) {
-                System.err.printf("Connection failed - retrying in %d second(s).\n", sleep / 1000);
-                try { Thread.sleep(sleep); } catch (Exception interruted) {}
-                if (sleep < 8000) sleep += sleep;
-            }
-        }
-        System.out.printf("Connected to VoltDB node at: %s.\n", server);
-    }
-
-    /**
-     * Connect to a set of servers in parallel. Each will retry until
-     * connection. This call will block until all have connected.
-     *
-     * @param servers A comma separated list of servers using the hostname:port
-     * syntax (where :port is optional).
-     * @throws InterruptedException if anything bad happens with the threads.
-     */
-    void connect(String servers) throws InterruptedException {
-        System.out.println("Connecting to VoltDB...");
-
-        String[] serverArray = servers.split(",");
-        final CountDownLatch connections = new CountDownLatch(serverArray.length);
-
-        // use a new thread to connect to each server
-        for (final String server : serverArray) {
-            new Thread(new Runnable() {
-                @Override
-                public void run() {
-                    connectToOneServerWithRetry(server);
-                    connections.countDown();
-                }
-            }).start();
-        }
-        // block until all have connected
-        connections.await();
-    }
-
-    /**
-     * Create a Timer task to display performance data on the Vote procedure
-     * It calls printStatistics() every displayInterval seconds
-     */
-    public void schedulePeriodicStats() {
-        timer = new Timer();
-        TimerTask statsPrinting = new TimerTask() {
-            @Override
-            public void run() { printStatistics(); }
-        };
-        timer.scheduleAtFixedRate(statsPrinting,
-                                  config.displayinterval * 1000,
-                                  config.displayinterval * 1000);
-    }
-
-    /**
-     * Prints a one line update on performance that can be printed
-     * periodically during a benchmark.
-     */
-    public synchronized void printStatistics() {
-        ClientStats stats = periodicStatsContext.fetchAndResetBaseline().getStats();
-        long time = Math.round((stats.getEndTimestamp() - benchmarkStartTS) / 1000.0);
-
-        System.out.printf("%02d:%02d:%02d ", time / 3600, (time / 60) % 60, time % 60);
-        System.out.printf("Throughput %d/s, ", stats.getTxnThroughput());
-        System.out.printf("Aborts/Failures %d/%d, ",
-                stats.getInvocationAborts(), stats.getInvocationErrors());
-
-        // cast to stats.getAverageLatency from long to double
-        System.out.printf("Avg/95%% Latency %.2f/%dms\n",
-                          stats.getAverageLatency(),
-                          stats.kPercentileLatency(0.95));
-
-    }
-
-    /**
-     * Prints the results of the voting simulation and statistics
-     * about performance.
-     *
-     * @throws Exception if anything unexpected happens.
-     */
-    public synchronized void printResults() throws Exception {
-        printHeading("Transaction Results");
-    BenchmarkCallback.printAllResults();
-
-        ClientStats stats = fullStatsContext.fetch().getStats();
-
-        // 3. Performance statistics
-        printHeading("Client Workload Statistics");
-
-        System.out.printf("Average throughput:            %,9d txns/sec\n", stats.getTxnThroughput());
-        System.out.printf("Average latency:               %,9.2f ms\n", stats.getAverageLatency());
-        System.out.printf("95th percentile latency:       %,9d ms\n", stats.kPercentileLatency(.95));
-        System.out.printf("99th percentile latency:       %,9d ms\n", stats.kPercentileLatency(.99));
-
-        printHeading("System Server Statistics");
-
-        if (config.autotune) {
-            System.out.printf("Targeted Internal Avg Latency: %,9d ms\n", config.latencytarget);
-        }
-        System.out.printf("Reported Internal Avg Latency: %,9.2f ms\n", stats.getAverageInternalLatency());
-
-        // 4. Write stats to file if requested
-        client.writeSummaryCSV(stats, config.statsfile);
-    }
-
-
-    /**
-     * Initializes any data that needs to be loaded into VoltDB before running the benchmark
-     * Override this method for customization
-     * @throws Exception
-     */
-    public void initialize() throws Exception {
-    }
-
-    /**
-     * Performs one iteration of the benchmark
-     * Override this method for customization
-     * @throws Exception
-     */
-    public void iterate() throws Exception {
-    }
-
-    /**
-     * Core benchmark code.
-     * Connect. Initialize. Run the loop. Cleanup. Print Results.
-     *
-     * @throws Exception if anything unexpected happens.
-     */
-    public void runBenchmark() throws Exception {
-        printHeading("Setup & Initialization");
-
-        // connect to one or more servers, loop until success
-        connect(config.servers);
-
-        // initialize using synchronous call
-        System.out.println("\nPre-loading Tables...\n");
-        initialize();
-
-        // ensure all data from initialize is committed before proceeding to iterate().
-        client.drain();
-
-        // Run the benchmark loop for the requested warmup time
-        // The throughput may be throttled depending on client configuration
-        System.out.println("Warming up for the specified "+ config.warmup +" seconds...");
-        final long warmupEndTime = System.currentTimeMillis() + (1000l * config.warmup);
-        while (warmupEndTime > System.currentTimeMillis()) {
-            iterate();
-        }
-
-        printHeading("Starting Benchmark");
-
-        // reset the stats after warmup
-        fullStatsContext.fetchAndResetBaseline();
-        periodicStatsContext.fetchAndResetBaseline();
-
-        // print periodic statistics to the console
-        benchmarkStartTS = System.currentTimeMillis();
-        schedulePeriodicStats();
-
-        // Run the benchmark loop for the requested duration
-        // The throughput may be throttled depending on client configuration
-        System.out.println("\nRunning benchmark...");
-        final long benchmarkEndTime = System.currentTimeMillis() + (1000l * config.duration);
-        while (benchmarkEndTime > System.currentTimeMillis()) {
-            iterate();
-        }
-
-        // cancel periodic stats printing
-        timer.cancel();
-
-        // block until all outstanding txns return
-        client.drain();
-
-        // print the summary results
-        printResults();
-
-        // close down the client connections
-        client.close();
-    }
-
-    /**
-     * Prints headings
-     */
-    public static void printHeading(String heading) {
-        System.out.print("\n"+HORIZONTAL_RULE);
-        System.out.println(" " + heading);
-        System.out.println(HORIZONTAL_RULE);
-    }
-
-}
