+++ examples/bank-offers/client/bankoffers/OfferBenchmark.java
@@ -27,26 +27,323 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Random;
+import java.util.Timer;
+import java.util.TimerTask;
+import java.util.concurrent.CountDownLatch;
 
+import org.voltdb.CLIConfig;
+import org.voltdb.client.Client;
+import org.voltdb.client.ClientConfig;
+import org.voltdb.client.ClientFactory;
+import org.voltdb.client.ClientResponse;
+import org.voltdb.client.ClientStats;
+import org.voltdb.client.ClientStatsContext;
+import org.voltdb.client.ClientStatusListenerExt;
+import org.voltdb.client.ProcedureCallback;
 import org.voltdb.types.TimestampType;
 
-public class OfferBenchmark extends BaseBenchmark {
+import com.google_voltpatches.common.collect.ConcurrentHashMultiset;
+import com.google_voltpatches.common.collect.Multiset;
+
+public class OfferBenchmark {
+
+    // handy, rather than typing this out several times
+    public static final String HORIZONTAL_RULE =
+            "----------" + "----------" + "----------" + "----------" +
+            "----------" + "----------" + "----------" + "----------" + "\n";
+
+    // validated command line configuration
+    final BankOffersConfig config;
+    // Reference to the database connection we will use
+    final Client client;
+    // Timer for periodic stats printing
+    Timer timer;
+    // Benchmark start time
+    long benchmarkStartTS;
+    // Statistics manager objects from the client
+    final ClientStatsContext periodicStatsContext;
+    final ClientStatsContext fullStatsContext;
 
     private Random rand = new Random();
     private long txnId = 0;
     private Long[] accounts;
     private String[] acct_states;
-    private int[] amounts = {25,50,75,100,150,200,250,300};
+    private int[] amounts = { 25, 50, 75, 100, 150, 200, 250, 300 };
     private PersonGenerator gen = new PersonGenerator();
-    private String[] offers = {"$5 off any purchase over $25","20% off any purchase over $50","Extra 25% off sale items"};
+    private String[] offers = {
+            "$5 off any purchase over $25",
+            "20% off any purchase over $50",
+            "Extra 25% off sale items" };
+
+    /**
+     * Prints headings
+     */
+    public static void printHeading(String heading) {
+        System.out.print("\n"+HORIZONTAL_RULE);
+        System.out.println(" " + heading);
+        System.out.println(HORIZONTAL_RULE);
+    }
+
+    /**
+     * Uses CLIConfig class to declaratively state command line options
+     * with defaults and validation.
+     */
+    public static class BankOffersConfig extends CLIConfig {
+
+        // STANDARD BENCHMARK OPTIONS
+        @Option(desc = "Comma separated list of the form server[:port] to connect to.")
+        String servers = "localhost";
+
+        @Option(desc = "Volt user name")
+        public String user = "";
+
+        @Option(desc = "Volt password")
+        public String password = "";
+
+        @Option(desc = "Benchmark duration, in seconds.")
+        int duration = 20;
+
+        @Option(desc = "Interval for performance feedback, in seconds.")
+        long displayinterval = 5;
+
+        @Option(desc = "Warmup duration in seconds.")
+        int warmup = 2;
+
+        @Option(desc = "Maximum TPS rate for benchmark.")
+        int ratelimit = 100000;
+
+        @Option(desc = "Determine transaction rate dynamically based on latency.")
+        boolean autotune = true;
+
+        @Option(desc = "Server-side latency target for auto-tuning.")
+        int latencytarget = 6;
+
+        @Option(desc = "Filename to write raw summary statistics to.")
+        String statsfile = "";
+
+        // CUSTOM OPTIONS
+        @Option(desc = "Number of customers to generate")
+        int custcount = 100000;
+
+        @Option(desc = "Number of vendors to generate")
+        int vendorcount = 5000;
+
+        @Override
+        public void validate() {
+            if (duration <= 0) exitWithMessageAndUsage("duration must be > 0");
+            if (warmup < 0) exitWithMessageAndUsage("warmup must be >= 0");
+            if (displayinterval <= 0) exitWithMessageAndUsage("displayinterval must be > 0");
+            if (ratelimit <= 0) exitWithMessageAndUsage("ratelimit must be > 0");
+            if (latencytarget <= 0) exitWithMessageAndUsage("latencytarget must be > 0");
+        }
+    }
 
     // constructor
-    public OfferBenchmark(BenchmarkConfig config) {
-        super(config);
+    public OfferBenchmark(BankOffersConfig config) {
+        this.config = config;
+
+        ClientConfig clientConfig = new ClientConfig(config.user, config.password, new StatusListener());
+        if (config.autotune) {
+            clientConfig.enableAutoTune();
+            clientConfig.setAutoTuneTargetInternalLatency(config.latencytarget);
+        }
+        else {
+            clientConfig.setMaxTransactionsPerSecond(config.ratelimit);
+        }
+        client = ClientFactory.createClient(clientConfig);
+
+        periodicStatsContext = client.createStatsContext();
+        fullStatsContext = client.createStatsContext();
+
+        printHeading("Command Line Configuration");
+        System.out.println(config.getConfigDumpString());
+    }
+
+    /**
+     * Provides a callback to be notified on node failure.
+     * This example only logs the event.
+     */
+    class StatusListener extends ClientStatusListenerExt {
+        @Override
+        public void connectionLost(String hostname, int port, int connectionsLeft, DisconnectCause cause) {
+            // if the benchmark is still active
+            if ((System.currentTimeMillis() - benchmarkStartTS) < (config.duration * 1000)) {
+                System.err.printf("Connection to %s:%d was lost.\n", hostname, port);
+            }
+        }
+    }
+
+    /**
+     * Connect to a single server with retry. Limited exponential backoff.
+     * No timeout. This will run until the process is killed if it's not
+     * able to connect.
+     *
+     * @param server hostname:port or just hostname (hostname can be ip).
+     */
+    void connectToOneServerWithRetry(String server) {
+        int sleep = 1000;
+        while (true) {
+            try {
+                client.createConnection(server);
+                break;
+            }
+            catch (Exception e) {
+                System.err.printf("Connection failed - retrying in %d second(s).\n", sleep / 1000);
+                try { Thread.sleep(sleep); } catch (Exception interruted) {}
+                if (sleep < 8000) sleep += sleep;
+            }
+        }
+        System.out.printf("Connected to VoltDB node at: %s.\n", server);
+    }
+
+    /**
+     * Connect to a set of servers in parallel. Each will retry until
+     * connection. This call will block until all have connected.
+     *
+     * @param servers A comma separated list of servers using the hostname:port
+     * syntax (where :port is optional).
+     * @throws InterruptedException if anything bad happens with the threads.
+     */
+    void connect(String servers) throws InterruptedException {
+        System.out.println("Connecting to VoltDB...");
+
+        String[] serverArray = servers.split(",");
+        final CountDownLatch connections = new CountDownLatch(serverArray.length);
+
+        // use a new thread to connect to each server
+        for (final String server : serverArray) {
+            new Thread(new Runnable() {
+                @Override
+                public void run() {
+                    connectToOneServerWithRetry(server);
+                    connections.countDown();
+                }
+            }).start();
+        }
+        // block until all have connected
+        connections.await();
+    }
+
+    /**
+     * Create a Timer task to display performance data on the Vote procedure
+     * It calls printStatistics() every displayInterval seconds
+     */
+    public void schedulePeriodicStats() {
+        timer = new Timer();
+        TimerTask statsPrinting = new TimerTask() {
+            @Override
+            public void run() { printStatistics(); }
+        };
+        timer.scheduleAtFixedRate(statsPrinting,
+                                  config.displayinterval * 1000,
+                                  config.displayinterval * 1000);
+    }
+
+    /**
+     * Prints a one line update on performance that can be printed
+     * periodically during a benchmark.
+     */
+    public synchronized void printStatistics() {
+        ClientStats stats = periodicStatsContext.fetchAndResetBaseline().getStats();
+        long time = Math.round((stats.getEndTimestamp() - benchmarkStartTS) / 1000.0);
+
+        System.out.printf("%02d:%02d:%02d ", time / 3600, (time / 60) % 60, time % 60);
+        System.out.printf("Throughput %d/s, ", stats.getTxnThroughput());
+        System.out.printf("Aborts/Failures %d/%d, ",
+                stats.getInvocationAborts(), stats.getInvocationErrors());
+
+        // cast to stats.getAverageLatency from long to double
+        System.out.printf("Avg/95%% Latency %.2f/%dms\n",
+                          stats.getAverageLatency(),
+                          stats.kPercentileLatency(0.95));
+
+    }
+
+    /**
+     * Prints the results of the voting simulation and statistics
+     * about performance.
+     *
+     * @throws Exception if anything unexpected happens.
+     */
+    public synchronized void printResults() throws Exception {
+        printHeading("Transaction Results");
+        BenchmarkCallback.printAllResults();
+
+        ClientStats stats = fullStatsContext.fetch().getStats();
+
+        // 3. Performance statistics
+        printHeading("Client Workload Statistics");
+
+        System.out.printf("Average throughput:            %,9d txns/sec\n", stats.getTxnThroughput());
+        System.out.printf("Average latency:               %,9.2f ms\n", stats.getAverageLatency());
+        System.out.printf("95th percentile latency:       %,9d ms\n", stats.kPercentileLatency(.95));
+        System.out.printf("99th percentile latency:       %,9d ms\n", stats.kPercentileLatency(.99));
+
+        printHeading("System Server Statistics");
+
+        if (config.autotune) {
+            System.out.printf("Targeted Internal Avg Latency: %,9d ms\n", config.latencytarget);
+        }
+        System.out.printf("Reported Internal Avg Latency: %,9.2f ms\n", stats.getAverageInternalLatency());
+
+        // 4. Write stats to file if requested
+        client.writeSummaryCSV(stats, config.statsfile);
+    }
+
+    public static class BenchmarkCallback implements ProcedureCallback {
+
+        private static Multiset<String> calls = ConcurrentHashMultiset.create();
+        private static Multiset<String> commits = ConcurrentHashMultiset.create();
+        private static Multiset<String> rollbacks = ConcurrentHashMultiset.create();
+
+        String procedureName;
+        long maxErrors;
+
+        public static void printProcedureResults(String procedureName) {
+            System.out.println("  " + procedureName);
+            System.out.println("        calls: " + calls.count(procedureName));
+            System.out.println("      commits: " + commits.count(procedureName));
+            System.out.println("    rollbacks: " + rollbacks.count(procedureName));
+        }
+
+        public static void printAllResults() {
+        for (String e : calls.elementSet()) {
+            printProcedureResults(e);
+        }
+        }
+
+        public BenchmarkCallback(String procedure, long maxErrors) {
+            super();
+            this.procedureName = procedure;
+            this.maxErrors = maxErrors;
+        }
+
+        public BenchmarkCallback(String procedure) {
+            this(procedure, 5l);
+        }
+
+        @Override
+        public void clientCallback(ClientResponse cr) {
+
+        calls.add(procedureName,1);
+
+            if (cr.getStatus() == ClientResponse.SUCCESS) {
+            commits.add(procedureName,1);
+            } else {
+                long totalErrors = rollbacks.add(procedureName,1);
+
+                System.err.println("DATABASE ERROR: " + cr.getStatusString());
+
+                if (totalErrors > maxErrors) {
+                    System.err.println("exceeded " + maxErrors + " maximum database errors - exiting client");
+                    System.exit(-1);
+                }
+
+            }
+        }
     }
 
     // this gets run once before the benchmark begins
-    @Override
     public void initialize() throws Exception {
 
         List<Long> acctList = new ArrayList<Long>(config.custcount*2);
@@ -116,7 +413,6 @@ public class OfferBenchmark extends BaseBenchmark {
         }
     }
 
-    @Override
     public void iterate() throws Exception {
 
         // pick a random account and generate a transaction
@@ -131,7 +427,7 @@ public class OfferBenchmark extends BaseBenchmark {
         // generate "out of state" fraud
         // a small % of the time, use a random state
         if (rand.nextInt(50000) == 0) {
-            txnState = gen.randomState();
+            txnState = PersonGenerator.randomState();
         }
 
         client.callProcedure(new BenchmarkCallback("CheckForOffers"),
@@ -140,10 +436,66 @@ public class OfferBenchmark extends BaseBenchmark {
 
     }
 
-    public static void main(String[] args) throws Exception {
-        BenchmarkConfig config = BenchmarkConfig.getConfig("OfferBenchmark",args);
+    /**
+     * Core benchmark code.
+     * Connect. Initialize. Run the loop. Cleanup. Print Results.
+     *
+     * @throws Exception if anything unexpected happens.
+     */
+    public void runBenchmark() throws Exception {
+        printHeading("Setup & Initialization");
 
-        BaseBenchmark c = new OfferBenchmark(config);
+        // connect to one or more servers, loop until success
+        connect(config.servers);
+
+        // initialize using synchronous call
+        System.out.println("\nPre-loading Tables...\n");
+        initialize();
+
+        // Run the benchmark loop for the requested warmup time
+        // The throughput may be throttled depending on client configuration
+        System.out.println("Warming up for the specified "+ config.warmup +" seconds...");
+        final long warmupEndTime = System.currentTimeMillis() + (1000l * config.warmup);
+        while (warmupEndTime > System.currentTimeMillis()) {
+            iterate();
+        }
+
+        printHeading("Starting Benchmark");
+
+        // reset the stats after warmup
+        fullStatsContext.fetchAndResetBaseline();
+        periodicStatsContext.fetchAndResetBaseline();
+
+        // print periodic statistics to the console
+        benchmarkStartTS = System.currentTimeMillis();
+        schedulePeriodicStats();
+
+        // Run the benchmark loop for the requested duration
+        // The throughput may be throttled depending on client configuration
+        System.out.println("\nRunning benchmark...");
+        final long benchmarkEndTime = System.currentTimeMillis() + (1000l * config.duration);
+        while (benchmarkEndTime > System.currentTimeMillis()) {
+            iterate();
+        }
+
+        // cancel periodic stats printing
+        timer.cancel();
+
+        // block until all outstanding txns return
+        client.drain();
+
+        // print the summary results
+        printResults();
+
+        // close down the client connections
+        client.close();
+    }
+
+    public static void main(String[] args) throws Exception {
+        BankOffersConfig config = new BankOffersConfig();
+        config.parse(OfferBenchmark.class.getName(), args);
+
+        OfferBenchmark c = new OfferBenchmark(config);
         c.runBenchmark();
     }
